# Spike on exercising ICS20 transfers to the Osmosis testnet with Hermes