--- a/crates/core/component/ibc/src/component/action_handler/ibc_action.rs
+++ b/crates/core/component/ibc/src/component/action_handler/ibc_action.rs
@@ -1,6 +1,6 @@
 use std::sync::Arc;
 
-use anyhow::Result;
+use anyhow::{Context, Result};
 use async_trait::async_trait;
 use penumbra_component::ActionHandler;
 use penumbra_storage::{StateRead, StateWrite};
@@ -42,21 +42,66 @@ impl ActionHandler for IbcAction {
 
     async fn execute<S: StateWrite>(&self, state: S) -> Result<()> {
         match self {
-            IbcAction::CreateClient(msg) => msg.try_execute(state).await?,
-            IbcAction::UpdateClient(msg) => msg.try_execute(state).await?,
-            IbcAction::ConnectionOpenInit(msg) => msg.try_execute(state).await?,
-            IbcAction::ConnectionOpenTry(msg) => msg.try_execute(state).await?,
-            IbcAction::ConnectionOpenAck(msg) => msg.try_execute(state).await?,
-            IbcAction::ConnectionOpenConfirm(msg) => msg.try_execute(state).await?,
-            IbcAction::ChannelOpenInit(msg) => msg.try_execute(state).await?,
-            IbcAction::ChannelOpenTry(msg) => msg.try_execute(state).await?,
-            IbcAction::ChannelOpenAck(msg) => msg.try_execute(state).await?,
-            IbcAction::ChannelOpenConfirm(msg) => msg.try_execute(state).await?,
-            IbcAction::ChannelCloseInit(msg) => msg.try_execute(state).await?,
-            IbcAction::ChannelCloseConfirm(msg) => msg.try_execute(state).await?,
-            IbcAction::RecvPacket(msg) => msg.try_execute(state).await?,
-            IbcAction::Acknowledgement(msg) => msg.try_execute(state).await?,
-            IbcAction::Timeout(msg) => msg.try_execute(state).await?,
+            IbcAction::CreateClient(msg) => msg
+                .try_execute(state)
+                .await
+                .with_context(|| "Failed to execute CreateClient message")?,
+            IbcAction::UpdateClient(msg) => msg
+                .try_execute(state)
+                .await
+                .with_context(|| "Failed to execute UpdateClient message")?,
+            IbcAction::ConnectionOpenInit(msg) => msg
+                .try_execute(state)
+                .await
+                .with_context(|| "Failed to execute ConnectionOpenInit message")?,
+            IbcAction::ConnectionOpenTry(msg) => msg
+                .try_execute(state)
+                .await
+                .with_context(|| "Failed to execute ConnectionOpenTry message")?,
+            IbcAction::ConnectionOpenAck(msg) => msg
+                .try_execute(state)
+                .await
+                .with_context(|| "Failed to execute ConnectionOpenAck message")?,
+            IbcAction::ConnectionOpenConfirm(msg) => msg
+                .try_execute(state)
+                .await
+                .with_context(|| "Failed to execute ConnectionOpenConfirm message")?,
+            IbcAction::ChannelOpenInit(msg) => msg
+                .try_execute(state)
+                .await
+                .with_context(|| "Failed to execute ChannelOpenInit message")?,
+            IbcAction::ChannelOpenTry(msg) => msg
+                .try_execute(state)
+                .await
+                .with_context(|| "Failed to execute ChannelOpenTry message")?,
+            IbcAction::ChannelOpenAck(msg) => msg
+                .try_execute(state)
+                .await
+                .with_context(|| "Failed to execute ChannelOpenAck message")?,
+            IbcAction::ChannelOpenConfirm(msg) => msg
+                .try_execute(state)
+                .await
+                .with_context(|| "Failed to execute ChannelOpenConfirm message")?,
+            IbcAction::ChannelCloseInit(msg) => msg
+                .try_execute(state)
+                .await
+                .with_context(|| "Failed to execute ChannelCloseInit message")?,
+            IbcAction::ChannelCloseConfirm(msg) => msg
+                .try_execute(state)
+                .await
+                .with_context(|| "Failed to execute ChannelCloseConfirm message")?,
+            IbcAction::RecvPacket(msg) => msg
+                .try_execute(state)
+                .await
+                .with_context(|| "Failed to execute RecvPacket message")?,
+            IbcAction::Acknowledgement(msg) => msg
+                .try_execute(state)
+                .await
+                .with_context(|| "Failed to execute Acknowledgement message")?,
+            IbcAction::Timeout(msg) => msg
+                .try_execute(state)
+                .await
+                .with_context(|| "Failed to execute Timeout message")?,
             IbcAction::Unknown(msg) => {
                 anyhow::bail!("unknown IBC message type: {}", msg.type_url)
             }