--- a/crates/core/component/ibc/src/component/channel.rs
+++ b/crates/core/component/ibc/src/component/channel.rs
@@ -11,6 +11,8 @@ use ibc_types::core::channel::{ChannelEnd, ChannelId, Packet, PortId};
 use penumbra_proto::{StateReadProto, StateWriteProto};
 use penumbra_storage::{StateRead, StateWrite};
 
+// Note: many of the methods on this trait need to write raw bytes,
+// because the data they write is interpreted by counterparty chains.
 #[async_trait]
 pub trait StateWriteExt: StateWrite + StateReadExt {
     fn put_channel_counter(&mut self, counter: u64) {
@@ -32,15 +34,24 @@ pub trait StateWriteExt: StateWrite + StateReadExt {
     }
 
     fn put_ack_sequence(&mut self, channel_id: &ChannelId, port_id: &PortId, sequence: u64) {
-        self.put_proto::<u64>(SeqAckPath::new(port_id, channel_id).to_string(), sequence);
+        self.put_raw(
+            SeqAckPath::new(port_id, channel_id).to_string(),
+            sequence.to_be_bytes().to_vec(),
+        );
     }
 
     fn put_recv_sequence(&mut self, channel_id: &ChannelId, port_id: &PortId, sequence: u64) {
-        self.put_proto::<u64>(SeqRecvPath::new(port_id, channel_id).to_string(), sequence);
+        self.put_raw(
+            SeqRecvPath::new(port_id, channel_id).to_string(),
+            sequence.to_be_bytes().to_vec(),
+        );
     }
 
     fn put_send_sequence(&mut self, channel_id: &ChannelId, port_id: &PortId, sequence: u64) {
-        self.put_proto::<u64>(SeqSendPath::new(port_id, channel_id).to_string(), sequence);
+        self.put_raw(
+            SeqSendPath::new(port_id, channel_id).to_string(),
+            sequence.to_be_bytes().to_vec(),
+        );
     }
 
     fn put_packet_receipt(&mut self, packet: &Packet) {
@@ -55,9 +66,7 @@ pub trait StateWriteExt: StateWrite + StateReadExt {
             CommitmentPath::new(&packet.port_on_a, &packet.chan_on_a, packet.sequence).to_string();
         let packet_hash = commit_packet(packet);
 
-        // TODO: this is wrong, it's important this is _RAW BYTES_, not a proto encoding.
-        // need to audit the other state writes here
-        self.put_proto::<Vec<u8>>(commitment_key, packet_hash);
+        self.put_raw(commitment_key, packet_hash);
     }
 
     fn delete_packet_commitment(
@@ -66,7 +75,7 @@ pub trait StateWriteExt: StateWrite + StateReadExt {
         port_id: &PortId,
         sequence: u64,
     ) {
-        self.put_proto::<Vec<u8>>(
+        self.put_raw(
             CommitmentPath::new(port_id, channel_id, sequence.into()).to_string(),
             vec![],
         );
@@ -79,7 +88,7 @@ pub trait StateWriteExt: StateWrite + StateReadExt {
         sequence: u64,
         acknowledgement: &[u8],
     ) {
-        self.put_proto::<Vec<u8>>(
+        self.put_raw(
             AckPath::new(port_id, channel_id, sequence.into()).to_string(),
             commit_acknowledgement(acknowledgement),
         );
@@ -106,25 +115,58 @@ pub trait StateReadExt: StateRead {
     }
 
     async fn get_recv_sequence(&self, channel_id: &ChannelId, port_id: &PortId) -> Result<u64> {
-        self.get_proto::<u64>(&SeqRecvPath::new(port_id, channel_id).to_string())
-            .await
-            .map(|sequence| sequence.unwrap_or(0))
+        if let Some(be_bytes) = self
+            .get_raw(&SeqRecvPath::new(port_id, channel_id).to_string())
+            .await?
+        {
+            // Parse big endian bytes into u64
+            Ok(u64::from_be_bytes(
+                be_bytes
+                    .try_into()
+                    .expect("written values are always 8-byte arrays"),
+            ))
+        } else {
+            // Default value for no key
+            Ok(0)
+        }
     }
 
     async fn get_ack_sequence(&self, channel_id: &ChannelId, port_id: &PortId) -> Result<u64> {
-        self.get_proto::<u64>(&SeqAckPath::new(port_id, channel_id).to_string())
-            .await
-            .map(|sequence| sequence.unwrap_or(0))
+        if let Some(be_bytes) = self
+            .get_raw(&SeqAckPath::new(port_id, channel_id).to_string())
+            .await?
+        {
+            // Parse big endian bytes into u64
+            Ok(u64::from_be_bytes(
+                be_bytes
+                    .try_into()
+                    .expect("written values are always 8-byte arrays"),
+            ))
+        } else {
+            // Default value for no key
+            Ok(0)
+        }
     }
 
     async fn get_send_sequence(&self, channel_id: &ChannelId, port_id: &PortId) -> Result<u64> {
-        self.get_proto::<u64>(&SeqSendPath::new(port_id, channel_id).to_string())
-            .await
-            .map(|sequence| sequence.unwrap_or(0))
+        if let Some(be_bytes) = self
+            .get_raw(&SeqSendPath::new(port_id, channel_id).to_string())
+            .await?
+        {
+            // Parse big endian bytes into u64
+            Ok(u64::from_be_bytes(
+                be_bytes
+                    .try_into()
+                    .expect("written values are always 8-byte arrays"),
+            ))
+        } else {
+            // Default value for no key
+            Ok(0)
+        }
     }
 
     async fn seen_packet(&self, packet: &Packet) -> Result<bool> {
-        self.get_proto::<String>(
+        self.get_raw(
             &ReceiptPath::new(&packet.port_on_b, &packet.chan_on_b, packet.sequence).to_string(),
         )
         .await
@@ -137,17 +179,16 @@ pub trait StateReadExt: StateRead {
         port_id: &PortId,
         sequence: u64,
     ) -> Result<bool> {
-        self.get_proto::<String>(
-            &ReceiptPath::new(port_id, channel_id, sequence.into()).to_string(),
-        )
-        .await
-        .map(|res| res.filter(|s| !s.is_empty()))
-        .map(|res| res.is_some())
+        // TODO: make this logic more explicit
+        self.get_raw(&ReceiptPath::new(port_id, channel_id, sequence.into()).to_string())
+            .await
+            .map(|res| res.filter(|s| !s.is_empty()))
+            .map(|res| res.is_some())
     }
 
     async fn get_packet_commitment(&self, packet: &Packet) -> Result<Option<Vec<u8>>> {
         let commitment = self
-            .get_proto::<Vec<u8>>(
+            .get_raw(
                 &CommitmentPath::new(&packet.port_on_a, &packet.chan_on_a, packet.sequence.into())
                     .to_string(),
             )
@@ -170,9 +211,7 @@ pub trait StateReadExt: StateRead {
         sequence: u64,
     ) -> Result<Option<Vec<u8>>> {
         let commitment = self
-            .get_proto::<Vec<u8>>(
-                &CommitmentPath::new(port_id, channel_id, sequence.into()).to_string(),
-            )
+            .get_raw(&CommitmentPath::new(port_id, channel_id, sequence.into()).to_string())
             .await?;
 
         // this is for the special case where the commitment is empty, we consider this None.
@@ -191,7 +230,7 @@ pub trait StateReadExt: StateRead {
         channel_id: &ChannelId,
         sequence: u64,
     ) -> Result<Option<Vec<u8>>> {
-        self.get_proto::<Vec<u8>>(&AckPath::new(port_id, channel_id, sequence.into()).to_string())
+        self.get_raw(&AckPath::new(port_id, channel_id, sequence.into()).to_string())
             .await
     }
 }