--- a/crates/core/component/ibc/src/component/client_counter.rs
+++ b/crates/core/component/ibc/src/component/client_counter.rs
@@ -224,7 +224,12 @@ pub fn validate_penumbra_client_state(
 
     // check client proof specs match penumbra proof specs
     if PENUMBRA_PROOF_SPECS.clone() != tm_client_state.proof_specs {
-        anyhow::bail!("invalid client state: proof specs do not match");
+        // allow legacy proof specs without prehash_key_before_comparison
+        let mut spec_with_prehash_key = tm_client_state.proof_specs.clone();
+        spec_with_prehash_key[0].prehash_key_before_comparison = true;
+        if PENUMBRA_PROOF_SPECS.clone() != spec_with_prehash_key {
+            anyhow::bail!("invalid client state: proof specs do not match");
+        }
     }
 
     // check that the trust level is correct