… Cargo.lock
… crates/core/component/ibc/Cargo.toml
… crates/core/component/ibc/src/component/transfer.rs
… crates/core/component/ibc/src/ics20_withdrawal.rs
… crates/proto/src/gen/penumbra.core.ibc.v1alpha1.rs
… crates/proto/src/gen/penumbra.core.ibc.v1alpha1.serde.rs
… crates/proto/src/state/read.rs
… crates/util/tendermint-proxy/src/tendermint_proxy.rs
… crates/util/tower-trace/src/v034/request_ext.rs
… deployments/relayer/configs/osmosis-testnet.json
… proto/go/gen/penumbra/core/ibc/v1alpha1/ibc.pb.go
… proto/penumbra/penumbra/core/ibc/v1alpha1/ibc.proto