Includes #3036

The goal of this PR, together with https://github.com/penumbra-zone/hermes/pull/1 , is to successfully exercise an ICS20 transfer between a Penumbra devnet and the Osmosis testnet, making any necessary changes to Penumbra and Hermes as needed along the way.

Current status (update as needed, details in comments below):

- 😬 https://github.com/penumbra-zone/penumbra/issues/3045
- ✅ Clients/Connections/Channels: With a temporary hack to allow incorrect proof specs that won't work with noninclusion proofs, we can successfully create a channel, e.g., [this one](https://testnet.mintscan.io/osmosis-testnet/txs/9EB7A84602DFA060BE16069BC12A2E42BD0D275C8B184D97EC1DF6F54567F913?height=2696909)
- ✅ ICS20 transfers Penumbra => Osmosis.
- ✅ ICS20 transfers Osmosis => Penumbra.