@hdevalence 2023-09-17
Issues we discovered so far:

- https://github.com/penumbra-zone/penumbra/issues/3045
- The definition of `timeout_height` we used in our own `Ics20Withdrawal` message was incorrect, and only allowed specifying the block height but not the revision number. Fixed.
- #3042 . Fixed, potentially: the relevant `ibc-go` code is here https://github.com/cosmos/ibc-go/blob/242cb9f630269ae53c92345d0b98ae5fcb6be6b7/modules/apps/transfer/types/packet.go#L72C24-L72C24 and also calls `MustSortJSON`; it's unclear if the sorting is actually required for counterparty chains.
- Our IBC packet ack GRPC query errored on un-acked packets instead of indicating that they were not acked. Fixed.
- Our packet handler rejected packets with empty timeout timestamps, instead of interpreting them as "no timeout". Fixed.