@hdevalence 2023-09-17
Current steps to reproduce the (evolving) test setup:

### Create a Penumbra devnet off this branch

With this branch checked out, create a new Penumbra devnet:
```
~/c/p/penumbra2 (hermes-osmo-test-5-spike|✔) [1]$ cargo run --release --bin pd -- testnet unsafe-reset-all
    Finished release [optimized] target(s) in 0.38s
     Running `target/release/pd testnet unsafe-reset-all`
~/c/p/penumbra2 (hermes-osmo-test-5-spike|✔) $ cargo run --release --bin pd -- testnet generate
    Finished release [optimized] target(s) in 0.35s
     Running `target/release/pd testnet generate`
2023-09-17T19:24:29.350882Z  INFO pd: Generating network config chain_id="penumbra-testnet-tethys-8777cb20"
2023-09-17T19:24:29.612580Z  INFO pd: Writing config files for network n_validators=1 chain_id=penumbra-testnet-tethys-8777cb20
2023-09-17T19:24:29.614315Z  INFO pd::testnet::config: Writing validator configs to config_dir=/Users/hdevalence/.penumbra/testnet_data/node0
```

Make note of the chain ID. It's also helpful to set an appropriate tracing filter (this is using `fish` syntax, adjust appropriately)
```
set -x RUST_LOG "info,[CheckTx]=debug"
set -x RUST_LOG "info,[CheckTx]=debug,[Query]=debug"
```
The `[]` syntax filters on spans, so the first does `info` level filtering for everything, and drops to `debug` only for events occurring during a `CheckTx`.  This is a generally useful trick for working with Penumbra testnets.

Start `pd` and `tendermint`:
```
~/c/p/penumbra2 (hermes-osmo-test-5-spike|✔) $ cargo run --release --bin pd -- start --home ~/.penumbra/testnet_data/node0/pd
```
```
~ $ tendermint --home ~/.penumbra/testnet_data/node0/tendermint/ start
```

### Configure Hermes

First, check out the Hermes branch in the draft PR: https://github.com/penumbra-zone/hermes/pull/1
Then build Hermes:
```
cargo build --release
```

The Hermes branch has a working-copy config file, `config-devnet-osmosis.toml`.

**Edit it so that the ID of the Penumbra chain `penumbra-testnet-tethys-RANDOMHEX` matches the ID of the newly running chain.** We'll use this config file in what follows.

Next, add keys for Osmosis. Create a text file with the mnemonic in it (cf https://github.com/informalsystems/hermes/issues/3625). It's probably useful to use the mnemonic for `osmo1kh0fwkdy05yp579d8vczgharkcexfw582zj488`, which is in 1Password.  Otherwise, make a new osmosis address and get funds from their testnet faucet.
```
cat "SEED PHRASE" > ./mnemonic
cargo run --release --bin hermes -- --config config-devnet-osmosis.toml keys add --chain osmo-test-5 --mnemonic-file ./mnemonic
```
Adding keys for Penumbra is not required, because the Penumbra backend doesn't currently sign transactions.  This may cause Hermes to report a supposedly fatal error on launch, but the error is not fatal.

If Hermes is configured to use that account, all of the transactions will be visible on the Osmosis testnet explorer on its account page: https://testnet.mintscan.io/osmosis-testnet/account/osmo1kh0fwkdy05yp579d8vczgharkcexfw582zj488

### Create a new transfer channel

Use Hermes to create a new channel between your local devnet and the Osmosis testnet:
```
cargo run --release --bin hermes -- --config config-devnet-osmosis.toml create channel --a-chain penumbra-testnet-tethys-8777cb20 --b-chain osmo-test-5 --a-port transfer --b-port transfer --new-client-connection
```
This will prompt for confirmation, then build a channel. This part should only need to be done once per devnet and should work:
```
     Running `target/release/hermes --config config-devnet-osmosis.toml create channel --a-chain penumbra-testnet-tethys-8777cb20 --b-chain osmo-test-5 --a-port transfer --b-port transfer --new-client-connection`
2023-09-17T19:47:24.723191Z  INFO ThreadId(01) running Hermes v1.6.0+6113050e-dirty
WARN: Are you sure you want a new connection & clients to be created? Hermes will use default security parameters.
Hint: Consider using the default invocation

hermes create channel --a-port <PORT-ID> --b-port <PORT-ID> --a-chain <CHAIN-A-ID> --a-connection <CONNECTION-A-ID>

WARN: Are you sure you want a new connection & clients to be created? Hermes will use default security parameters.
Hint: Consider using the default invocation

hermes create channel --a-port <PORT-ID> --b-port <PORT-ID> --a-chain <CHAIN-A-ID> --a-connection <CONNECTION-A-ID>

to re-use a pre-existing connection. yes
2023-09-17T19:47:32.410997Z  INFO ThreadId(01) Creating new clients, new connection, and a new channel with order ORDER_UNORDERED
2023-09-17T19:47:32.412533Z  WARN ThreadId(25) foreign_client.create{client=osmo-test-5->penumbra-testnet-tethys-8777cb20:07-tendermint-0}: global telemetry state not set, will initialize it using default histogram ranges
2023-09-17T19:47:49.294865Z  INFO ThreadId(01) foreign_client.create{client=osmo-test-5->penumbra-testnet-tethys-8777cb20:07-tendermint-0}: 🍭 client was created successfully id=07-tendermint-1
2023-09-17T19:47:50.130564Z  WARN ThreadId(25) foreign_client.create{client=penumbra-testnet-tethys-8777cb20->osmo-test-5:07-tendermint-0}:send_messages_and_wait_commit{chain=osmo-test-5 tracking_id=create client}:send_tx_with_account_sequence_retry{chain=osmo-test-5 account.sequence=54}:estimate_gas: tx simulation successful but no gas amount used was returned, falling back on default gas: 400000
2023-09-17T19:47:55.695076Z  INFO ThreadId(01) foreign_client.create{client=penumbra-testnet-tethys-8777cb20->osmo-test-5:07-tendermint-0}: 🍭 client was created successfully id=07-tendermint-1029
2023-09-17T19:48:09.765655Z  INFO ThreadId(01) 🥂 penumbra-testnet-tethys-8777cb20 => OpenInitConnection(OpenInit { Attributes { connection_id: connection-0, client_id: 07-tendermint-1, counterparty_connection_id: None, counterparty_client_id: 07-tendermint-1029 } }) at height 0-224
2023-09-17T19:48:31.400128Z  WARN ThreadId(25) send_messages_and_wait_commit{chain=osmo-test-5 tracking_id=ConnectionOpenTry}:send_tx_with_account_sequence_retry{chain=osmo-test-5 account.sequence=55}:estimate_gas: tx simulation successful but no gas amount used was returned, falling back on default gas: 400000
2023-09-17T19:48:36.419668Z  INFO ThreadId(01) 🥂 osmo-test-5 => OpenTryConnection(OpenTry { Attributes { connection_id: connection-939, client_id: 07-tendermint-1029, counterparty_connection_id: connection-0, counterparty_client_id: 07-tendermint-1 } }) at height 5-2720518
2023-09-17T19:48:41.777273Z  WARN ThreadId(25) send_messages_and_wait_commit{chain=osmo-test-5 tracking_id=update client on source for ConnectionOpenAck}:send_tx_with_account_sequence_retry{chain=osmo-test-5 account.sequence=56}:estimate_gas: tx simulation successful but no gas amount used was returned, falling back on default gas: 400000
2023-09-17T19:49:05.955246Z  INFO ThreadId(01) 🥂 penumbra-testnet-tethys-8777cb20 => OpenAckConnection(OpenAck { Attributes { connection_id: connection-0, client_id: 07-tendermint-1, counterparty_connection_id: connection-939, counterparty_client_id: 07-tendermint-1029 } }) at height 0-235
2023-09-17T19:49:11.884281Z  WARN ThreadId(25) send_messages_and_wait_commit{chain=osmo-test-5 tracking_id=ConnectionOpenConfirm}:send_tx_with_account_sequence_retry{chain=osmo-test-5 account.sequence=57}:estimate_gas: tx simulation successful but no gas amount used was returned, falling back on default gas: 400000
2023-09-17T19:49:13.993327Z  INFO ThreadId(01) 🥂 osmo-test-5 => OpenConfirmConnection(OpenConfirm { Attributes { connection_id: connection-939, client_id: 07-tendermint-1029, counterparty_connection_id: connection-0, counterparty_client_id: 07-tendermint-1 } }) at height 5-2720525
2023-09-17T19:49:17.806922Z  INFO ThreadId(01) connection handshake already finished for Connection { delay_period: 0ns, a_side: ConnectionSide { chain: BaseChainHandle { chain_id: penumbra-testnet-tethys-8777cb20 }, client_id: 07-tendermint-1, connection_id: connection-0 }, b_side: ConnectionSide { chain: BaseChainHandle { chain_id: osmo-test-5 }, client_id: 07-tendermint-1029, connection_id: connection-939 } }
2023-09-17T19:49:31.493034Z  INFO ThreadId(01) 🎊  penumbra-testnet-tethys-8777cb20 => OpenInitChannel(OpenInit { port_id: transfer, channel_id: channel-0, connection_id: None, counterparty_port_id: transfer, counterparty_channel_id: None }) at height 0-240
2023-09-17T19:49:37.511024Z  WARN ThreadId(25) send_messages_and_wait_commit{chain=osmo-test-5 tracking_id=ChannelOpenTry}:send_tx_with_account_sequence_retry{chain=osmo-test-5 account.sequence=58}:estimate_gas: tx simulation successful but no gas amount used was returned, falling back on default gas: 400000
2023-09-17T19:49:40.152439Z  INFO ThreadId(01) 🎊  osmo-test-5 => OpenTryChannel(OpenTry { port_id: transfer, channel_id: channel-1675, connection_id: connection-939, counterparty_port_id: transfer, counterparty_channel_id: channel-0 }) at height 5-2720530
2023-09-17T19:50:02.174903Z  INFO ThreadId(01) 🎊  penumbra-testnet-tethys-8777cb20 => OpenAckChannel(OpenAck { port_id: transfer, channel_id: channel-0, connection_id: connection-0, counterparty_port_id: transfer, counterparty_channel_id: channel-1675 }) at height 0-246
2023-09-17T19:50:08.717664Z  WARN ThreadId(25) send_messages_and_wait_commit{chain=osmo-test-5 tracking_id=ChannelOpenConfirm}:send_tx_with_account_sequence_retry{chain=osmo-test-5 account.sequence=59}:estimate_gas: tx simulation successful but no gas amount used was returned, falling back on default gas: 400000
2023-09-17T19:50:13.730280Z  INFO ThreadId(01) 🎊  osmo-test-5 => OpenConfirmChannel(OpenConfirm { port_id: transfer, channel_id: channel-1675, connection_id: connection-939, counterparty_port_id: transfer, counterparty_channel_id: channel-0 }) at height 5-2720536
2023-09-17T19:50:17.445328Z  INFO ThreadId(01) channel handshake already finished for Channel { ordering: ORDER_UNORDERED, a_side: ChannelSide { chain: BaseChainHandle { chain_id: penumbra-testnet-tethys-8777cb20 }, client_id: 07-tendermint-1, connection_id: connection-0, port_id: transfer, channel_id: channel-0, version: None }, b_side: ChannelSide { chain: BaseChainHandle { chain_id: osmo-test-5 }, client_id: 07-tendermint-1029, connection_id: connection-939, port_id: transfer, channel_id: channel-1675, version: None }, connection_delay: 0ns }
SUCCESS Channel {
    ordering: Unordered,
    a_side: ChannelSide {
        chain: BaseChainHandle {
            chain_id: ChainId {
                id: "penumbra-testnet-tethys-8777cb20",
                version: 0,
            },
            runtime_sender: Sender { .. },
        },
        client_id: ClientId(
            "07-tendermint-1",
        ),
        connection_id: ConnectionId(
            "connection-0",
        ),
        port_id: PortId(
            "transfer",
        ),
        channel_id: Some(
            ChannelId(
                "channel-0",
            ),
        ),
        version: None,
    },
    b_side: ChannelSide {
        chain: BaseChainHandle {
            chain_id: ChainId {
                id: "osmo-test-5",
                version: 5,
            },
            runtime_sender: Sender { .. },
        },
        client_id: ClientId(
            "07-tendermint-1029",
        ),
        connection_id: ConnectionId(
            "connection-939",
        ),
        port_id: PortId(
            "transfer",
        ),
        channel_id: Some(
            ChannelId(
                "channel-1675",
            ),
        ),
        version: None,
    },
    connection_delay: 0ns,
}
```

Hermes will then exit.